feat: add expectation queries(#109)
