--- strawberry/src/pages/expectation/hooks/apis/useExpectationPageQuery.tsx
@@ -1,17 +1,0 @@
-import { useQuery } from "react-query";
-
-import { useFetch } from "../../../../data/config/useFetch";
-
-export function useExpectationPageQuery() {
-  const getExpectationPage = useFetch({
-    url: "expectation/land",
-    method: "GET",
-  });
-
-  const query = useQuery({
-    queryKey: ["expectationPage"],
-    queryFn: getExpectationPage,
-  });
-
-  return query;
-}
